Pluggable framebuffer implementations [WIP]
'diff_to_previous' will compare the previous frame and try and calculate
the differences and pass args to a device to only render those small
changes. This can be very quick for small screen updates, but suffers
from variable render times, depending on the changes applied.

'full_frame' always renders the full frame every time. This is slower
as there are more pixels to update on every render, but has a more
consistent render time.